added filter option based on customization functionalties will be implement later

+++ src/Pages/MyArtAndCraftList/MyArtAndCraftList.jsx
@@ -3,7 +3,8 @@ import { Helmet } from "react-helmet";
 import { useLoaderData } from "react-router-dom";
 import { AuthContext } from "../../AuthProvider/AuthProvider";
 import SingleCard from "./SingleCard";
-
+import Swal from "sweetalert2";
+import "./select.css";
 const MyArtAndCraftList = () => {
   const { user } = useContext(AuthContext);
   // console.log(user.email);
@@ -20,20 +21,64 @@ const MyArtAndCraftList = () => {
   );
   // console.log(availableArtAndCrafts);
   const handleDelete = (_id) => {
-    console.log(_id);
-    fetch(`http://localhost:5000/update/${_id}`, {
-      method: "DELETE",
-    })
-      .then((res) => res.json())
-      .then((data) => {
-        console.log(availableArtAndCrafts);
-        const remaining = availableArtAndCrafts.filter(
-          (matchedArtCraft) => matchedArtCraft._id !== _id
-        );
-        console.log(availableArtAndCrafts);
-        setAvailableArtAndCrafts(remaining);
-      });
+    // console.log(_id);
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        fetch(`http://localhost:5000/update/${_id}`, {
+          method: "DELETE",
+        })
+          .then((res) => res.json())
+          .then((data) => {
+            if (data.deletedCount > 0) {
+              Swal.fire({
+                title: "Success!",
+                text: "Art/Craft has been Deleted!",
+                icon: "success",
+                confirmButtonText: "Delete another",
+              });
+              const remaining = availableArtAndCrafts.filter(
+                (matchedArtCraft) => matchedArtCraft._id !== _id
+              );
+              // console.log(availableArtAndCrafts);
+              setAvailableArtAndCrafts(remaining);
+            }
+          });
+      }
+    });
   };
+  // const handleBookFilter = (filter) => {
+  //   if (filter === "default") {
+  //     setDisplayReadBook(readBook);
+  //   } else if (filter === "totalPages") {
+  //     const sortedBooks = sortBooksByRatingDescending(
+  //       [...readBook],
+  //       "totalPages"
+  //     );
+  //     setDisplayReadBook(sortedBooks);
+  //   } else if (filter === "rating") {
+  //     const sortedBooks = sortBooksByRatingDescending([...readBook], "rating");
+  //     setDisplayReadBook(sortedBooks);
+  //   } else if (filter === "yearOfPublishing") {
+  //     const sortedBooks = sortBooksByRatingDescending(
+  //       [...readBook],
+  //       "yearOfPublishing"
+  //     );
+  //     setDisplayReadBook(sortedBooks);
+  //   }
+  // };
+  // const getOptionValue = () => {
+  //   let optionValue = document.getElementById("list").value;
+
+  //   handleBookFilter(optionValue);
+  // };
 
   return (
     <div className="w-[90%] mx-auto">
@@ -45,7 +90,18 @@ const MyArtAndCraftList = () => {
       <h1 className="text-5xl text-white font-bold mt-6 mb-10 text-center">
         Your Personal Collection
       </h1>
-      <div className="grid grid-cols-3 lg:gap-10">
+      <div className="flex justify-center">
+        <div className="box  " id="list">
+          <select>
+            <option>Option 1</option>
+            <option>Option 2</option>
+            <option>Option 3</option>
+            <option>Option 4</option>
+            <option>Option 5</option>
+          </select>
+        </div>
+      </div>
+      <div className="grid grid-cols-1 md:grid-cols-2 md:gap-5 lg:grid-cols-3 lg:gap-10 mt-16">
         {availableArtAndCrafts.map((card) => (
           <SingleCard
             key={card._id}
